Deprecate utils/ dir in favor of domain-scoped modules
- Absorb all subtitle utils into koffee/subtitle.py
- Inline get_video_duration into koffee/asr.py as _get_video_duration
- Promote retry.py to koffee/_retry.py at the package root
- Delete koffee/utils/ entirely
- Update all imports and mock patch paths in tests
