+++ tests/test_retry.py
@@ -3,7 +3,7 @@
 import pytest
 from pytest_mock import MockerFixture
 
-from koffee.utils import with_retries
+from koffee._retry import with_retries
 
 
 class RetryableError(Exception):
@@ -34,7 +34,7 @@ def test_with_retries_returns_on_first_success() -> None:
 
 def test_with_retries_retries_on_retryable_error(mocker: MockerFixture) -> None:
     """Tests that a retryable error triggers a retry and then succeeds."""
-    mocker.patch("koffee.utils.retry.time.sleep")
+    mocker.patch("koffee._retry.time.sleep")
     attempts = {"n": 0}
 
     def fn():
@@ -65,7 +65,7 @@ def test_with_retries_propagates_non_retryable_immediately() -> None:
 
 def test_with_retries_exhaustion_raises_last_error(mocker: MockerFixture) -> None:
     """Tests that after max_retries, the last retryable error is re-raised."""
-    mocker.patch("koffee.utils.retry.time.sleep")
+    mocker.patch("koffee._retry.time.sleep")
 
     def fn():
         raise RetryableError("still flaky")
@@ -76,7 +76,7 @@ def test_with_retries_exhaustion_raises_last_error(mocker: MockerFixture) -> Non
 
 def test_with_retries_uses_exponential_backoff(mocker: MockerFixture) -> None:
     """Tests that backoff doubles between retries."""
-    mock_sleep = mocker.patch("koffee.utils.retry.time.sleep")
+    mock_sleep = mocker.patch("koffee._retry.time.sleep")
 
     def fn():
         raise RetryableError("flaky")
